+++ src/rogallo/data/state/ui.py
@@ -38,6 +38,9 @@ class UIState:
     with_link_jumps: bool = True
     """Should the application support jumping to links via numeric labels?"""
 
+    cosy_link_jumps: bool = False
+    """Should the numeric labels be displayed in a cosy way?"""
+
 
 ##############################################################################
 def ui_state_file() -> Path: