+++ ChangeLog.md
@@ -38,6 +38,9 @@
 - **BREAKING CHANGE:** Removed the `ToggleLinkNumbers` command. This is now
   controlled by a key binding specific to the viewer widget.
   ([#453](https://github.com/davep/rogallo/pull/453))
+- **BREAKING CHANGE:** Removed the `ToggleCosyLinkNumbers` command. This is
+  now controlled by a key binding specific to the viewer widget.
+  ([#453](https://github.com/davep/rogallo/pull/453))
 
 ## v2.4.0